requirements.txt
tweepy
Pillow
requests
zbar
